Create ad_skipper.py
Everytime this program re-opens through the main.py, we need to enter a auth code again (bcs of the process from 20 to 30. Maybe it can get automatized with the refresh token you can recive from the spotify api or you can automatize, that the code in the google link gets translated directly in the program, but I didnt figure it out so fast. So this would be your turn. ^^
Maybe you noticed the time sleeps for several seconds, I built them in, bcs sometimes it's to fast for spotify, but maybe there are more efficient solutions.